Makes the sync code optional. Using Re-Exporting to get backwards compatibility
