+++ rmp/src/sync/decode/str.rs
