+++ rmp/src/sync/encode/ext.rs
