+++ rmp/src/sync/encode/mod.rs
@@ -1,51 +1,27 @@
 //! Provides various functions and structs for MessagePack encoding.
 
-mod bin;
-mod dec;
-mod ext;
-mod map;
-mod sint;
-mod str;
-mod uint;
-mod vec;
+pub(crate) mod bin;
+pub(crate) mod dec;
+pub(crate) mod ext;
+pub(crate) mod map;
+pub(crate) mod sint;
+pub(crate) mod str;
+pub(crate) mod uint;
+pub(crate) mod vec;
 
-pub use self::bin::{write_bin, write_bin_len};
-pub use self::dec::{write_f32, write_f64};
-pub use self::sint::{write_i16, write_i32, write_i64, write_i8, write_nfix, write_sint};
-pub use self::str::{write_str, write_str_len};
-pub use self::uint::{write_pfix, write_u16, write_u32, write_u64, write_u8, write_uint};
-
-#[cfg(feature = "std")]
-use std::error;
-use core::fmt::{self, Display, Debug, Formatter};
 
 use crate::Marker;
 
 pub mod buffer;
+
 pub use buffer::ByteBuf;
+pub use crate::encode::{DataWriteError, ValueWriteError,RmpWriteErr};
+use crate::encode::{MarkerWriteError};
 
 #[doc(inline)]
 #[allow(deprecated)]
 pub use crate::errors::Error;
 
-/// The error type for operations on the [RmpWrite] trait.
-///
-/// For [std::io::Write], this is [std::io::Error]
-/// For [ByteBuf], this is [core::convert::Infallible]
-pub trait RmpWriteErr: Display + Debug + crate::errors::MaybeErrBound + 'static {}
-#[cfg(feature = "std")]
-impl RmpWriteErr for std::io::Error {}
-impl RmpWriteErr for core::convert::Infallible {}
-
-// An error returned from the `write_marker` and `write_fixval` functions.
-struct MarkerWriteError<E: RmpWriteErr>(E);
-
-impl<E: RmpWriteErr> From<E> for MarkerWriteError<E> {
-    #[cold]
-    fn from(err: E) -> Self {
-        MarkerWriteError(err)
-    }
-}
 
 
 /// Attempts to write the given marker into the writer.
@@ -53,17 +29,6 @@ fn write_marker<W: RmpWrite>(wr: &mut W, marker: Marker) -> Result<(), MarkerWri
     wr.write_u8(marker.to_u8()).map_err(MarkerWriteError)
 }
 
-/// An error returned from primitive values write functions.
-#[doc(hidden)]
-pub struct DataWriteError<E: RmpWriteErr>(E);
-
-impl<E: RmpWriteErr> From<E> for DataWriteError<E> {
-    #[cold]
-    #[inline]
-    fn from(err: E) -> DataWriteError<E> {
-        DataWriteError(err)
-    }
-}
 
 
 /// Encodes and attempts to write a nil value into the given write.
@@ -104,14 +69,18 @@ pub fn write_bool<W: RmpWrite>(wr: &mut W, val: bool) -> Result<(), W::Error> {
     write_marker(wr, marker).map_err(|e| e.0)
 }
 
-mod sealed{
+mod sealed {
     pub trait Sealed {}
+
     #[cfg(feature = "std")]
     impl<T: ?Sized + std::io::Write> Sealed for T {}
+
     #[cfg(not(feature = "std"))]
     impl Sealed for &mut [u8] {}
+
     #[cfg(not(feature = "std"))]
     impl Sealed for alloc::vec::Vec<u8> {}
+
     impl Sealed for super::ByteBuf {}
 }
 
@@ -194,62 +163,6 @@ impl<T: std::io::Write> RmpWrite for T {
     }
 }
 
-/// An error that can occur when attempting to write multi-byte MessagePack value.
-#[derive(Debug)]
-#[allow(deprecated)] // TODO: Needed for compatibility
-pub enum ValueWriteError<E: RmpWriteErr = Error> {
-    /// I/O error while writing marker.
-    InvalidMarkerWrite(E),
-    /// I/O error while writing data.
-    InvalidDataWrite(E),
-}
-
-impl<E: RmpWriteErr> From<MarkerWriteError<E>> for ValueWriteError<E> {
-    #[cold]
-    fn from(err: MarkerWriteError<E>) -> Self {
-        match err {
-            MarkerWriteError(err) => ValueWriteError::InvalidMarkerWrite(err),
-        }
-    }
-}
-
-impl<E: RmpWriteErr> From<DataWriteError<E>> for ValueWriteError<E> {
-    #[cold]
-    fn from(err: DataWriteError<E>) -> Self {
-        match err {
-            DataWriteError(err) => ValueWriteError::InvalidDataWrite(err),
-        }
-    }
-}
-
-#[cfg(feature = "std")] // Backwards compatbility ;)
-impl From<ValueWriteError<std::io::Error>> for std::io::Error {
-    #[cold]
-    fn from(err: ValueWriteError<std::io::Error>) -> std::io::Error {
-        match err {
-            ValueWriteError::InvalidMarkerWrite(err) |
-            ValueWriteError::InvalidDataWrite(err) => err,
-        }
-    }
-}
-
-#[cfg(feature = "std")]
-impl<E: RmpWriteErr> error::Error for ValueWriteError<E> {
-    #[cold]
-    fn source(&self) -> Option<&(dyn error::Error + 'static)> {
-        match *self {
-            ValueWriteError::InvalidMarkerWrite(ref err) |
-            ValueWriteError::InvalidDataWrite(ref err) => Some(err),
-        }
-    }
-}
-
-impl<E: RmpWriteErr> Display for ValueWriteError<E> {
-    #[cold]
-    fn fmt(&self, f: &mut Formatter<'_>) -> Result<(), fmt::Error> {
-        f.write_str("error while writing multi-byte MessagePack value")
-    }
-}
 
 /// Encodes and attempts to write the most efficient array length implementation to the given write,
 /// returning the marker used.
