+++ rmp/src/sync/encode/str.rs
