+++ rmp/src/sync/encode/vec.rs
