+++ rmp/src/sync/mod.rs
@@ -1,0 +1,3 @@
+pub(crate) mod encode;
+pub(crate) mod decode;
+
